lab3/pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = "User management system with PostgreSQL integration"
authors = ["namwoam <namwoam@gmail.com>"]
packages = [{include = "*.py"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
